mcp-gateway/requirements.txt
fastapi
uvicorn[standard]
httpx
psycopg2-binary
